Update to module elasticdata equipement
