+++ .gitignore
@@ -1,1 +1,6 @@
 /nb-configuration.xml
+/target/
+.project
+.classpath
+.settings/
+
