--- src/java/fr/paris/lutece/plugins/elasticdata/modules/gru/resources/dansmarue_messages.properties
@@ -1,3 +1,0 @@
-# Plugin's keys
-module.description=Module de production de donn\u00e9es ElasticSearch Kibana pour DansMaRue
-module.provider=Mairie de Paris
