+++ src/java/fr/paris/lutece/plugins/elasticdata/modules/gru/resources/equipement_messages.properties
@@ -1,0 +1,3 @@
+# Plugin's keys
+module.description=Module de production de donn\u00e9es ElasticSearch Kibana pour DansMaRue-Equipement
+module.provider=Mairie de Paris
