--- webapp/WEB-INF/conf/plugins/elasticdata-dansmarue.properties
@@ -1,7 +1,0 @@
-# elasticdata-dansmarue properties file
-
-elasticdata-dansmarue.canal.S=Web Service Paris.fr
-elasticdata-dansmarue.canal.B=Back Office 3975
-elasticdata-dansmarue.canal.A=Apple iOS
-elasticdata-dansmarue.canal.G=Google Android
-elasticdata-dansmarue.canal.unkown=Inconnu ou invalide
