+++ webapp/WEB-INF/conf/plugins/elasticdata-equipement.properties
@@ -1,0 +1,7 @@
+# elasticdata-equipement properties file
+
+elasticdata-equipement.canal.S=Web Service Paris.fr
+elasticdata-equipement.canal.B=Back Office 3975
+elasticdata-equipement.canal.A=Apple iOS
+elasticdata-equipement.canal.G=Google Android
+elasticdata-equipement.canal.unkown=Inconnu ou invalide
