+++ webapp/WEB-INF/conf/plugins/elasticdata-equipement_context.xml
@@ -5,7 +5,7 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd" >
        
        
-       <bean id="elasticdata-dansmarue.demandDataSource" class="fr.paris.lutece.plugins.elasticdata.modules.dansmarue.business.SignalementDataSource" >
+       <bean id="elasticdata-dansmarue.demandDataSource" class="fr.paris.lutece.plugins.elasticdata.modules.equipement.business.SignalementDataSource" >
            <property name="id" value="DemandDataSource" />
            <property name="name" value="Base des signalements DansMaRue" />
            <property name="targetIndexName" value="signalements" />
